Created admin user section Added privileges depending on role Added admin middleware Added default role to new users Added 403 error case Removed unnecessary js route guards

+++ app/Http/Controllers/Admin/RoleController.php
@@ -2,8 +2,8 @@
 
 namespace Blog\Http\Controllers\Admin;
 
-use Blog\Repositories\RoleRepository;
 use Blog\Http\Controllers\Controller;
+use Blog\Repositories\RoleRepository;
 use Illuminate\Http\Request;
 use Blog\RoleResourceCollection;
 use Blog\RoleResource;
@@ -28,7 +28,12 @@ class RoleController extends Controller
     {
         $params = $request->all() + $roleRepo->params();
 
-        $roles = new RoleResourceCollection($roleRepo->roles($locale, $params)->paginate(10));
+        $roles = $roleRepo->roles($locale, $params);
+        if ($params['page']) {
+            $roles = new RoleResourceCollection($roles->paginate(10));
+        } else {
+            $roles = new RoleResourceCollection($roles->get());
+        }
 
         return $roles->additional(compact('params'));
     }
@@ -40,6 +45,8 @@ class RoleController extends Controller
 
     public function store(Request $request)
     {
+        abort_unless(in_array($request->user()->role->id, [1, 3]), 403);
+
         $rules = $this->rules;
         $rules['name_en'][] = 'unique:roles,name_en';
         $rules['name_lv'][] = 'unique:roles,name_lv';
@@ -50,6 +57,8 @@ class RoleController extends Controller
 
     public function update(Request $request, $locale, Role $role)
     {
+        abort_unless(in_array($request->user()->role->id, [1, 3]), 403);
+
         $rules = $this->rules;
         $rules['name_en'][] = "unique:roles,name_en,{$role->id}";
         $rules['name_lv'][] = "unique:roles,name_lv,{$role->id}";
@@ -62,6 +71,8 @@ class RoleController extends Controller
 
     public function destroy($locale, Role $role)
     {
+        abort_unless(in_array($request->user()->role->id, [1, 3]), 403);
+
         $role->delete();
 
         return response()->json();
